xmlscraper/requirements.txt
scrapy==2.6.1
feedparser==6.0.8
Flask==2.1.1
psycopg2==2.9.3
python-dateutil==2.8.2

